Change how builtins are added
Builtins are now defined in a set of <foo>.builtin files which is
parsed by newt-builtins.py to generate all of the definitions. All
that the C code needs to have is the C functions themselves, which are
named newt_builtin_<name>, where <name> is the builtin name, with '_'
replacing '.'.

Signed-off-by: Keith Packard <keithp@keithp.com>