Fix in list iterator. Report string bounds errors.
Sharing the string/list fetch functions caused errors at the end of
the tuple due to the intentional out of bounds access. Add a
'report_error' to both list and string fetch functions so that errors
can be suppressed in the 'in list' case and also now report string
bounds errors in other cases.

Signed-off-by: Keith Packard <keithp@keithp.com>