Track parsing status to allow separate middle-of-statement prompts
Both the posix and arduino versions now print different prompts based
on whether the parser is waiting for a new statement or not. This
makes it easier to understand when extra newlines are needed to 'make it go'.

Also made lexing non-interactive content skip over blank lines just
like comments so that the program can have whitespace.

Signed-off-by: Keith Packard <keithp@keithp.com>