Don't generate -inf (0xff800000) as value of list at offset 0.
With snek_list == 0, snek_offset_to_poly(0, snek_list) looks like:

	(snek_poly_t) { .u = SNEK_EXPONENT_MASK | offset | type };

This results in a .u value of 0xff800000 (SNEK_EXPONENT_MASK), which
is the IEEE encoding for -inf.

Fixing this was 'interesting. The first change is to re-order
snek_type_t so that snek_builtin is encoding 0. Because snek_builtins
start at 1 instead of 0, there is no valid snek_builtin with all of
the IEEE mantissa bits set to zero. snek_builtin is the only
non-reference type, and the range is expected to be quite limited.

However, the code still confuses an offset of 0 with a NULL
'pointer'. To fix this, I've switched the offset representation of
NULL from 0 to ~3 -- that is, a maximum offset value with both type
bits cleared. That ends up looking like a builtin, which makes dealing
with this value in the memory code easy -- it follows the same path as
builtins, which are not references.

To keep this small enough for Arduino, I've moved some NULL checks in
memory management out of the individual types and down into the common
memory manager code.

Reported-by: fanoush
Signed-off-by: Keith Packard <keithp@keithp.com>