Disable -Warray-bounds when accessing _snek_mems
Because snek_builtin is type 0, this array would need an unused entry
were it indexed directly by type. Instead, it's indexed by type -
1. To avoid actually doing that computation at runtime, we do
(&_snek_mems[-1])[type], which makes the linker compute the base
address and then indexing by type 'just works'. However, the compiler
is not happy with the -1 index, generating -Warray-bounds warnings
when optimized. Suppress that warnings around this usage by placing
the usage in a static inline, instead of a macro.

Signed-off-by: Keith Packard <keithp@keithp.com>