Ignore incorrect compiler warning
In some code paths, when inlining functions in snek-memory.c, the
compiler gets confused into thinking that there might be an array
bounds error in pool_addr from SNEK_OFFSET_NONE sneaking in. This
"can't" happen (and is checked when SNEK_DEBUG is enabled), but the
compiler gets confused into thinking that it might and generates a
warning at this line of code. Disable the warning as I can't think of
any code change that would make this example pass the compiler and not
generate a lot more code.

Signed-off-by: Keith Packard <keithp@keithp.com>