metro-snek: Fix ADC reading
Reduce ADC clock to 93750Hz. This is well below the required 2.1MHz,
but appears to improve accuracy.

Disable CORREN bit -- I thought this was using the calibration data,
but it turns out to use correction factors presumably gathered from
elsewhere and stored in the gaincorr and offsetcorr registers.

Add various calls to ao_adc_sync. The docs seem to indicate that the
chip will dtrt by stalling the processor until the clock domains are
synchronized, but doing it manually seems reasonable?

Turn off pull-up/pull-down when in ADC mode.

Signed-off-by: Keith Packard <keithp@keithp.com>