doc: Add Makefile fragments to generate print-ready book files
These fragments combine the pages from snek-book.pdf in a way that
when printed double-sided (flip long edge), they can be folded into
a book with the pages in the correct order. A cover is generated that
fits on a regular 8.5 x 11 sheet so that it can also be printed.

Signed-off-by: Keith Packard <keithp@keithp.com>