snek-duino/itsybitsy-snek: Disable error checking when programming fuse bytes
The ATmega32u4 seems to return random values for the high four bits of
the efuse byte; according to the docs, those aren't used. This makes
verification fail, which causes any subsequent operations to be
skipped. Using -V -c makes avrdude skip verification of those
bytes. We don't want to skip verification of the firmware though, so
that's now done in a separtate step.

Signed-off-by: Keith Packard <keithp@keithp.com>