Require in-scope definition of op= targets
When performing an operator assignment (e.g. +=), a variable on the
LHS must have been declared in-scope either by creating a local
variable or defining the name as global.

Fortunately, the 'snek_op_is_local' function does just the check we
need -- this looks for the name in the local frame, which is true
either if the name references a local variable, or if the name is
marked global.

We just need to check for the op= case, and whether there is a local
frame in effect.

Signed-off-by: Keith Packard <keithp@keithp.com>