Clean up 'make install'
Stop installing snekboard .elf file

Make sure 'make install' builds everything necessary so that
a preceding 'make all' is not required.

Add dependency for doc/coderay-asciidoctor.css so that Make knows when
it will be built.

Install snek-riscv and snek-arm scripts to let users run them if they
like (they're in $(SHAREDIR), so not on the usual PATH).

Signed-off-by: Keith Packard <keithp@keithp.com>