Share common formal assignment code in snek_func_push
Most of the code for assigning formals is the same for named and
positional actuals. Create a single loop over the actuals and split
out the piece which figures out the name, either from the named
actual, or from the list of formals by position.

Also, add a pile of comments to this file.

Signed-off-by: Keith Packard <keithp@keithp.com>