chips/atmega: Improve timing precision
Instead of counting something approximating milliseconds, count raw
timer over flow values and mix in the current timer value to generate
a clock at the full precision of the timer, which runs at F_CPU/64.

On a 16MHz device, this offers a resolution of 4µs.

This also shares the time code between the duemilanove and mega ports.

Reported-by: Helmut Grohne <helmut@subdivi.de>
Signed-off-by: Keith Packard <keithp@keithp.com>