chips/samd21: Turn off built-in neopixels at init
This makes sure any built-in neopixels are turned off when the board
starts up, in case they were left on across reset, or if the chips
don't power up in an 'off' state (sigh). This means that any GPIOs
with built-in neopixels are going to get some traffic at startup.

Signed-off-by: Keith Packard <keithp@keithp.com>