ports/uno: Add port to Arduino Uno
The Uno looks a lot like the Duemilanove, but in place of the FT232
usb/serial converter chip it uses an ATmega16u2 (just like the Arduino
Mega2560 board). This suffers from the same lack of XON/XOFF flow
control support which is necessary for mu-editor to download programs.

So, as on the Mega board, we provide new USB serial converter firmware
and then run Snek at a lower baud rate which triggers the serial
converter firmware to enable xon/xoff flow control.

Signed-off-by: Keith Packard <keithp@keithp.com>