Handle 'chained' comparison operators
Python supports expressions of the form:

	a < b < c

where < can be replaced with any comparison operator. This is
converted into the natural conjunction:

	a < b && b < c

except that 'b' is only evaluated once.

To support this in snek, new op codes have been introduced which:

 1. Perform the comparison operation
 2a. If the resulting value is false
	set the accumulator to the false value
	branch to the offset specified as a parameter
 2b. If the resulting value is true
	leave the accumulator with the previous value
	skip over the offset in the instruction

When compiling, the original comparison opcode for all but the last
comparison are replaced with these new 'chain' comparison op codes.

Signed-off-by: Keith Packard <keithp@keithp.com>