Get rid of production used to work-around lola bug
Before version 1.4, lola had a bug when computing the follow set for
non-terminal that was repeated in a production. It would only compute
the values for the *first* instance of the non-terminal in the
production and not for subsequent instances. Hence this kludge which
split the dict-ent production in half so that 'expr' occurred only
once in each production. With that bug fixed, snek can now use a shorter
grammar and save some time and space while parsing.

Signed-off-by: Keith Packard <keithp@keithp.com>