Allow for builtin function aliases, alias 'pow' for math.pow
Python has a builtin 'pow' function, which performs the math.pow when
presented with two arguments (it does powmod with three). We can at least
make the two argument form work by having it be an alias for the existing
math.pow function on systems that include the math library.

Signed-off-by: Keith Packard <keithp@keithp.com>